requirements.txt
pyautogen[teachable]